+++ .eslintrc
@@ -19,6 +19,7 @@
     "react/no-unescaped-entities": "warn",
     "react/no-string-refs": "warn",
     "react/no-find-dom-node": "warn",
+    "no-undef":"error",
     "semi": ["error", "always", {"omitLastInOneLineBlock": true}],
     "no-console": "off"
   },
